add service discovery

+++ song-service/src/main/java/com/epam/learn/song/service/SongService.java
@@ -10,14 +10,19 @@ import com.epam.learn.song.model.SongMetadata;
 import com.epam.learn.song.repository.SongRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.transaction.Transactional;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.regex.Pattern;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class SongService {
+    private static final int MAX_ALLOWED_IDS_TO_DELETE = 200;
+    private static final Pattern VALID_ID_PATTERN = Pattern.compile("^\\d+$");
+
     private final ObjectMapper objectMapper;
     private final SongRepository songRepository;
 
@@ -31,7 +36,15 @@ public class SongService {
     }
 
     public SongMetadataResponse getSongById(String id) {
-        SongMetadata song = songRepository.findById(Integer.parseInt(id))
+        if (!VALID_ID_PATTERN.matcher(id).matches()) {
+            throw new ValidationException("Invalid ID format: ID should be a positive integer");
+        }
+
+        int idAsInteger = Integer.parseInt(id);
+        if (idAsInteger == 0) {
+            throw new ValidationException("Invalid ID format: ID should be a positive integer");
+        }
+        SongMetadata song = songRepository.findById(idAsInteger)
             .orElseThrow(() -> new NotFoundException("Song not found with id: " + id));
 
         return objectMapper.convertValue(song, SongMetadataResponse.class);
@@ -39,9 +52,16 @@ public class SongService {
 
     @Transactional
     public DeleteSongMetadataBulkResponse deleteSongs(String ids) {
-        List<Integer> idList = Arrays.stream(ids.split(","))
-            .map(Integer::parseInt)
-            .toList();
+        List<Integer> idList = new ArrayList<>();
+        try {
+            Arrays.stream(ids.split(",")).forEach(id -> idList.add(Integer.parseInt(id)));
+        } catch (NumberFormatException e) {
+            throw new ValidationException("Invalid ID format: IDs should be integers");
+        }
+
+        if (idList.size() > MAX_ALLOWED_IDS_TO_DELETE) {
+            throw new ValidationException("Number of ids should not exceed " + MAX_ALLOWED_IDS_TO_DELETE);
+        }
         songRepository.deleteAllById(idList);
         return new DeleteSongMetadataBulkResponse(idList);
     }
@@ -49,6 +69,14 @@ public class SongService {
     private void validateSong(CreateSongMetadataRequest song) throws ValidationException {
         if (song.name() == null || song.artist() == null) {
             throw new ValidationException("Song name and artist are required");
+        }
+
+        if (song.year() != null && !song.year().matches("^[12][0-9]{3}$")) {
+            throw new ValidationException("Year must be a four-digit number starting with 1 or 2");
+        }
+
+        if (song.length() != null && !song.length().matches("^\\d{2}:\\d{2}$")) {
+            throw new ValidationException("Length must be in the format MM:SS");
         }
     }
 }
